fix error
polynomial에 대한 최솟값과 방정식의 해를 구할 수 있다.

+++ Polynomial/queue.cpp
@@ -19,7 +19,7 @@ bool Queue::isEmpty()
 	return m_head == nullptr;
 }
 
-void Queue::enqueue(int value)
+void Queue::enqueue(double value)
 {
 	if (isEmpty()) {
 		m_head = new Node(value);
